[charts] Reuse shared date utils
