+++ packages/x-charts/src/internals/plugins/featurePlugins/useChartPolarAxis/computeAxisValue.ts
@@ -1,4 +1,4 @@
-import { scaleBand, scalePoint, scaleTime } from '@mui/x-charts-vendor/d3-scale';
+import { scaleBand, scalePoint } from '@mui/x-charts-vendor/d3-scale';
 import { AxisConfig, ScaleName } from '../../../../models';
 import {
   ChartsAxisProps,
@@ -14,6 +14,7 @@ import { ChartSeriesType, PolarChartSeriesType } from '../../../../models/series
 import { getColorScale, getOrdinalColorScale } from '../../../colorScale';
 import { getTickNumber, scaleTickNumberByRange } from '../../../ticks';
 import { getScale } from '../../../getScale';
+import { isDateData, createDateFormatter } from '../../../dateHelpers';
 import { getAxisExtremum } from './getAxisExtremum';
 import type { ChartDrawingArea } from '../../../../hooks';
 import { ChartSeriesConfig } from '../../models/seriesConfig';
@@ -53,18 +54,6 @@ function getRange(
     ];
   }
   return [0, Math.min(drawingArea.height, drawingArea.width) / 2];
-}
-
-const isDateData = (data?: readonly any[]): data is Date[] => data?.[0] instanceof Date;
-
-function createDateFormatter(
-  axis: AxisConfig<'band' | 'point', any, ChartsAxisProps>,
-  range: number[],
-): AxisConfig<'band' | 'point', any, ChartsAxisProps>['valueFormatter'] {
-  const timeScale = scaleTime(axis.data!, range);
-
-  return (v, { location }) =>
-    location === 'tick' ? timeScale.tickFormat(axis.tickNumber)(v) : `${v.toLocaleString()}`;
 }
 
 const DEFAULT_CATEGORY_GAP_RATIO = 0.2;
